fix(ci): python-coverage-shards — relocate websocket tests to keep part_001 under the limit
#657's new tests pushed tests/unit/api/test_observability_api_parts/test_observability_api_part_001.py
to 1593 lines, over the hard 1500-line first-party file limit checked in the coverage shards.
Move the TestWorkspaceWebSocket class plus its websocket-only helpers (_make_sync_test_client,
_make_empty_sync_test_client, _seed_websocket_workspace, _temporary_api_token, _RecordingWebSocket,
_FrameRecordingWebSocket) into a new test_observability_api_part_003.py. Shared scaffolding still
needed by both files (_clear_settings_cache, _auth, _create_workspace, _BODY/_REPOSITORY_BODY) is
duplicated. Every test name, body, parametrization, and assertion is preserved (part_001 1593 -> 637).
